% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COVID19MEXICO_EXAMPLE}
\alias{COVID19MEXICO_EXAMPLE}
\title{Some cases of covid as published by the ministry of health in Mexico}
\format{
A data frame with 53 rows and 35 variables:
\describe{
\item{FECHA_ACTUALIZACION}{Download and process the dictionary}
\item{ID_REGISTRO}{Download and process the dictionary}
\item{ORIGEN}{Download and process the dictionary}
\item{SECTOR}{Download and process the dictionary}
\item{ENTIDAD_UM}{Download and process the dictionary}
\item{SEXO}{Download and process the dictionary}
\item{ENTIDAD_NAC}{Download and process the dictionary}
\item{ENTIDAD_RES}{Download and process the dictionary}
\item{MUNICIPIO_RES}{Download and process the dictionary}
\item{TIPO_PACIENTE}{Download and process the dictionary}
\item{FECHA_INGRESO}{Download and process the dictionary}
\item{FECHA_SINTOMAS}{Download and process the dictionary}
\item{FECHA_DEF}{Download and process the dictionary}
\item{INTUBADO}{Download and process the dictionary}
\item{NEUMONIA}{Download and process the dictionary}
\item{EDAD}{Download and process the dictionary}
\item{NACIONALIDAD}{Download and process the dictionary}
\item{EMBARAZO}{Download and process the dictionary}
\item{HABLA_LENGUA_INDIG}{Download and process the dictionary}
\item{DIABETES}{Download and process the dictionary}
\item{EPOC}{Download and process the dictionary}
\item{ASMA}{Download and process the dictionary}
\item{INMUSUPR}{Download and process the dictionary}
\item{HIPERTENSION}{Download and process the dictionary}
\item{OTRA_COM}{Download and process the dictionary}
\item{CARDIOVASCULAR}{Download and process the dictionary}
\item{OBESIDAD}{Download and process the dictionary}
\item{RENAL_CRONICA}{Download and process the dictionary}
\item{TABAQUISMO}{Download and process the dictionary}
\item{OTRO_CASO}{Download and process the dictionary}
\item{RESULTADO}{Download and process the dictionary}
\item{MIGRANTE}{Download and process the dictionary}
\item{PAIS_NACIONALIDAD}{Download and process the dictionary}
\item{PAIS_ORIGEN}{Download and process the dictionary}
\item{UCI}{Download and process the dictionary}
}
}
\source{
\url{http://187.191.75.115/gobmx/salud/datos_abiertos/datos_abiertos_covid19.zip}
}
\usage{
COVID19MEXICO_EXAMPLE
}
\description{
A dataset containing the description of 53 cases of covid in Mexico
as released by the ministry of health
}
\details{
The description of variables is available downloading the dictionary
with the function \code{downloadDicHealth()} and then processing it
in variables in the local environment \code{variables.diccionario()}
}
\keyword{datasets}
