% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19mx}
\alias{covid19mx}
\title{covid19mx: Mexican COVID-19 Pandemic Data.}
\format{
A dataframe object with 3118237 rows and 40 variables:
\describe{
  \item{FECHA_ACTUALIZACION}{}
  \item{ID_REGISTRO}{}
  \item{ORIGEN}{}
  \item{SECTOR}{}
  \item{ENTIDAD_UM}{}
  \item{SEXO}{}
  \item{ENTIDAD_NAC}{}
  \item{ENTIDAD_RES}{}
  \item{MUNICIPIO_RES}{}
  \item{TIPO_PACIENTE}{}
  \item{FECHA_INGRESO}{}
  \item{FECHA_SINTOMAS}{}
  \item{FECHA_DEF}{}
  \item{INTUBADO}{}
  \item{NEUMONIA}{}
  \item{EDAD}{}
  \item{NACIONALIDAD}{}
  \item{EMBARAZO}{}
  \item{HABLA_LENGUA_INDIG}{}
  \item{INDIGENA}{}
  \item{DIABETES}{}
  \item{EPOC}{}
  \item{ASMA}{}
  \item{INMUSUPR}{}
  \item{HIPERTENSION}{}
  \item{OTRA_COM}{}
  \item{CARDIOVASCULAR}{}
  \item{OBESIDAD"}{}
  \item{RENAL_CRONICA}{}
  \item{TABAQUISMO}{}
  \item{OTRO_CASO}{}
  \item{TOMA_MUESTRA_LAB}{}
  \item{TOMA_MUESTRA_ANTIGENO}{}
  \item{RESULTADO_ANTIGENO}{}
  \item{CLASIFICACION_FINAL}{}
  \item{MIGRANTE}{}
  \item{PAIS_NACIONALIDAD}{}
  \item{PAIS_ORIGEN}{}
  \item{UCI}{}
  ...
}
}
\source{
\url{https://www.gob.mx/salud/documentos/datos-abiertos-152127?idiom=es}
}
\usage{
covid19mx
}
\description{
The covid19mx provides Mexican COVID-19 (SARS-COV2) Pandemic Data from the Secretary of Health <https://www.gob.mx/salud/documentos/datos-abiertos-152127?idiom=es>.
}
\details{
This the database is daily update by DGE.
}
\keyword{datasets}
