% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19mx20}
\alias{covid19mx20}
\title{covid19mx20: Mexican COVID-19 Pandemic Data 2020.}
\format{
A dataframe object with 3118237 rows and 42 variables:
\describe{
  \item{FECHA_ACTUALIZACION}{La base de datos se alimenta diariamente, esta variable permite identificar la fecha de la ultima actualizacion}
  \item{ID_REGISTRO}{Número identificador del caso}
  \item{ORIGEN}{La vigilancia centinela se realiza a través del sistema de unidades de salud monitoras de enfermedades respiratorias (USMER). Las USMER incluyen unidades médicas del primer, segundo o tercer nivel de atención y también participan como USMER las unidades de tercer nivel que por sus características contribuyen a ampliar el panorama de información epidemiológica, entre ellas las que cuenten con especialidad de neumología, infectología o pediatría. (Categorías en Catalógo Anexo).}
  \item{SECTOR}{Identifica el tipo de institución del Sistema Nacional de Salud que brindó la atención.}
  \item{ENTIDAD_UM}{Identifica la entidad donde se ubica la unidad medica que brindó la atención.}
  \item{SEXO}{Identifica al sexo del paciente.}
  \item{ENTIDAD_NAC}{Identifica la entidad de nacimiento del paciente.}
  \item{ENTIDAD_RES}{Identifica la entidad de residencia del paciente.}
  \item{MUNICIPIO_RES}{Identifica el municipio de residencia del paciente.}
  \item{TIPO_PACIENTE}{Identifica el tipo de atención que recibió el paciente en la unidad. Se denomina como ambulatorio si regresó a su casa o se denomina como hospitalizado si fue ingresado a hospitalización.}
  \item{FECHA_INGRESO}{Identifica la fecha de ingreso del paciente a la unidad de atención.}
  \item{FECHA_SINTOMAS}{Identifica la fecha en que inició la sintomatología del paciente.}
  \item{FECHA_DEF}{Identifica la fecha en que el paciente falleció.}
  \item{INTUBADO}{Identifica si el paciente requirió de intubación.}
  \item{NEUMONIA}{Identifica si al paciente se le diagnosticó con neumonía.}
  \item{EDAD}{Identifica la edad del paciente.}
  \item{NACIONALIDAD}{Identifica si el paciente es mexicano o extranjero.}
  \item{EMBARAZO}{Identifica si la paciente está embarazada.}
  \item{HABLA_LENGUA_INDIG}{Identifica si el paciente habla lengua índigena.}
  \item{INDIGENA}{Identifica si el paciente se autoidentifica como una persona indígena.}
  \item{DIABETES}{Identifica si el paciente tiene un diagnóstico de diabetes.}
  \item{EPOC}{Identifica si el paciente tiene un diagnóstico de EPOC.}
  \item{ASMA}{Identifica si el paciente tiene un diagnóstico de asma.}
  \item{INMUSUPR}{Identifica si el paciente presenta inmunosupresión.}
  \item{HIPERTENSION}{Identifica si el paciente tiene un diagnóstico de hipertensión.}
  \item{OTRA_COM}{Identifica si el paciente tiene diagnóstico de otras enfermedades.}
  \item{CARDIOVASCULAR}{Identifica si el paciente tiene un diagnóstico de enfermedades cardiovasculares.}
  \item{OBESIDAD"}{Identifica si el paciente tiene diagnóstico de obesidad.}
  \item{RENAL_CRONICA}{Identifica si el paciente tiene diagnóstico de insuficiencia renal crónica.}
  \item{TABAQUISMO}{Identifica si el paciente tiene hábito de tabaquismo.}
  \item{OTRO_CASO}{Identifica si el paciente tuvo contacto con algún otro caso diagnósticado con SARS CoV-2}
  \item{TOMA_MUESTRA_LAB}{"Identifica si al paciente se le tomó muestra de laboratorio.}
  \item{RESULTADO_LAB}{Identifica el resultado del análisis de la muestra reportado por el  laboratorio de la Red Nacional de Laboratorios de Vigilancia Epidemiológica (INDRE, LESP y LAVE) y laboratorios privados avalados por el InDRE cuyos resultados son registrados en SISVER. (Catálogo de resultados diagnósticos anexo).}
  \item{TOMA_MUESTRA_ANTIGENO}{Identifica si al paciente se le tomó muestra de antígeno para SARS-CoV-2}
  \item{RESULTADO_ANTIGENO}{Identifica el resultado del análisis de la muestra de antígeno tomada al paciente}
  \item{CLASIFICACION_FINAL}{Identifica si el paciente es un caso de COVID-19 según el catálogo \"CLASIFICACION_FINAL\".}
  \item{MIGRANTE}{Identifica si el paciente es una persona migrante.}
  \item{PAIS_NACIONALIDAD}{Identifica la nacionalidad del paciente.}
  \item{PAIS_ORIGEN}{Identifica el país del que partió el paciente rumbo a México.}
  \item{UCI}{Identifica si el paciente requirió ingresar a una Unidad de Cuidados Intensivos.}
  \item{date}{onset date.}
  \item{year}{year dataset.}
  ...
}
}
\source{
\url{https://www.gob.mx/salud/documentos/datos-abiertos-152127?idiom=es}
}
\usage{
covid19mx20
}
\description{
The covid19mx20 provides Mexican COVID-19 (SARS-COV2) Pandemic Data from the Secretary of Health <https://www.gob.mx/salud/documentos/datos-abiertos-152127?idiom=es>.
}
\details{
This the database is daily update by DGE.
}
\keyword{datasets}
