% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download.diccionario}
\alias{download.diccionario}
\title{Baja el diccionario de datos de la Secretaría de salud}
\usage{
download.diccionario()
}
\description{
Baja el diccionario de datos de la Secretaría de salud
}
\examples{
download.diccionario()
}
