% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downloadCovidNacionalFromCDMX}
\alias{downloadCovidNacionalFromCDMX}
\title{Downloads covid national cases from the CDMX site}
\usage{
downloadCovidNacionalFromCDMX(numRows = -1)
}
\arguments{
\item{numRows}{Numeric. Number of rows to download. Defaults to -1 for all}
}
\description{
Downloads covid national cases from the CDMX site
}
\examples{
downloadCovidNacionalFromCDMX(2)
}
