% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidMinistryHealth.R
\name{downloadDicHealth}
\alias{downloadDicHealth}
\title{Baja el diccionario de datos de la Secretaría de salud}
\usage{
downloadDicHealth()
}
\description{
Baja el diccionario de datos de la Secretaría de salud
}
\examples{
downloadDicHealth()
}
