% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downloadDictionarySINAVE}
\alias{downloadDictionarySINAVE}
\title{Downloads the dictinary of the SINAVE database}
\usage{
downloadDictionarySINAVE()
}
\description{
Baja el diccionario de datos de la Secretaría de salud desde el sitio de la CDMX
}
