% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downloadHospitalizedCdmxEdoMex}
\alias{downloadHospitalizedCdmxEdoMex}
\title{Downloads data with hospitalized persons in CDMX & EdoMex}
\usage{
downloadHospitalizedCdmxEdoMex(year = NULL, month = NULL)
}
\description{
Downloads data with hospitalized persons in CDMX & EdoMex
}
\examples{
downloadHospitalizedCdmxEdoMex(month = "enero", year = 2020)
}
