% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downloadHospitalsZmcm}
\alias{downloadHospitalsZmcm}
\title{Downloads data from hospitals in the metropolitan area of Mexico City}
\usage{
downloadHospitalsZmcm(year = NULL)
}
\description{
Downloads data from hospitals in the metropolitan area of Mexico City
}
\examples{
downloadHospitalsZmcm()
}
