% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downloadIPAS}
\alias{downloadIPAS}
\title{Downloads Inventario de Programas y Acciones Sociales ante el COVID-19}
\usage{
downloadIPAS(numRows = -1)
}
\arguments{
\item{numRows}{Numeric. Number of rows to download. Defaults to -1 for all}
}
\description{
https://datos.cdmx.gob.mx/explore/dataset/inventario-de-acciones-y-programas-sociales-rendicion-de-cuentas/information/
}
\details{
Fields:
nombre_programa_accion social
nombre_programa_accion_social
descripcion_programa_accion_social
tipo_apoyo_unidad
fecha_anuncio
unidad_responsable
categoria
fecha_publicacion_gaceta
enlace_a_gaceta
presupuesto_mxn
beneficiarios_poblacion_objetivo
metas_fisicas
metas_fisicas_unidad
}
