% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downloadPolicies}
\alias{downloadPolicies}
\title{Downloads governmental policies related to the covid19 epidemic}
\usage{
downloadPolicies(year = NULL)
}
\description{
Downloads governmental policies related to the covid19 epidemic
}
\examples{
downloadPolicies()
}
