% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downloadSINAVE}
\alias{downloadSINAVE}
\title{Downloads SINAVE database (suspicious cases)}
\usage{
downloadSINAVE()
}
\description{
Downloads SINAVE database (suspicious cases)
}
\examples{
downloadSINAVE()
}
