% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load.municipios}
\alias{load.municipios}
\title{Carga los datos geográficos de los municipios de México}
\usage{
load.municipios()
}
\description{
Carga los datos geográficos de los municipios de México
}
\examples{
load.municipios()
}
