% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocessIPAS}
\alias{preprocessIPAS}
\title{Preprocess Inventario de Programas y Acciones Sociales ante el COVID-19}
\usage{
preprocessIPAS()
}
\description{
Preprocess Inventario de Programas y Acciones Sociales ante el COVID-19
}
