% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit2.R
\name{strsplit2}
\alias{strsplit2}
\title{Function to split strings with the option to not remove the split}
\usage{
strsplit2(x, split, type = "remove", perl = FALSE, ...)
}
\arguments{
\item{x}{character vector}

\item{split}{character vector to use for splitting}

\item{type}{how to split, 'remove' to the split, 'before' to keep the split to the left and 'after' to keep the split to the right. Defaults to 'remove'.}

\item{perl}{logical. Should Perl-compatible regexps be used?}

\item{...}{other arguments to be passed to strsplit}
}
\description{
Function to split strings with the option to not remove the split
}
