% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{variables.diccionario}
\alias{variables.diccionario}
\title{Prepares catalogs in local variables}
\usage{
variables.diccionario()
}
\description{
Prepares catalogs in local variables
}
\examples{
variables.diccionario()
}
