% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downloadCovidNacionalFromCDMX}
\alias{downloadCovidNacionalFromCDMX}
\title{Downloads covid national cases from the CDMX site}
\usage{
downloadCovidNacionalFromCDMX()
}
\description{
Downloads covid national cases from the CDMX site
}
\examples{
downloadCovidNacionalFromCDMX()
}
