% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downloadDictionarySINAVE}
\alias{downloadDictionarySINAVE}
\title{Downloads the dictinary of the SINAVE database}
\usage{
downloadDictionarySINAVE()
}
\description{
Downloads the dictinary of the SINAVE database
}
