% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%>\%}
\alias{\%>\%}
\alias{download.covidmx}
\title{Baja los datos sobre covid de la Secretaría de salud}
\usage{
download.covidmx()
}
\description{
Baja los datos sobre covid de la Secretaría de salud
}
\examples{
download.covidmx()
}
