% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%>\%}
\alias{\%>\%}
\alias{download.covidmx}
\alias{data.frame.covidmx}
\title{Baja los datos sobre covid de la Secretaría de salud}
\usage{
download.covidmx()

data.frame.covidmx()
}
\description{
Baja los datos sobre covid de la Secretaría de salud

Prepares data frame of covidmx data
}
\examples{
download.covidmx()
data.frame.covidmx()
}
